requirements.txt
scholarly==1.7.11
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
unidecode>=1.3.0
